+++ test/test_alphastyle.jl
@@ -1,3 +1,4 @@
+using Test
 using OrderedCollections: OrderedDict
 import DocumenterCitations:
     _alpha_suffix,